+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Migrate to `security/ignorePageTypeRestriction` flag in TCA
 - Migrate and improve TCA configuration
+- Rename `Userfuncs` namespace to `UserFunctions`
 
 ### Removed
 - Remove obsolete TCEforms tag from FlexForms