+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 22-10-2024
+
 ### Added
 - Add support for nonce attribute output in `generateNonce` method
 - Add site set